% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_inat_data.R
\name{clean_inat_data}
\alias{clean_inat_data}
\title{Initial processing of iNaturalist project data}
\usage{
clean_inat_data(obsfile, npsboundary = NA)
}
\arguments{
\item{obsfile}{Full path to downloaded .csv file of the iNaturalist project}

\item{npsboundary}{filename with full path to NPS boundary shape file}
}
\value{
data frame containing cleaned iNaturalist data
}
\description{
Script to initiate data processing of iNaturalist project data
Uses downloaded .csv file of the iNaturalist project and nps_boundary shape file
Cleans the data to remove records from other years, missing geo-locations,
records outside parks
}
\examples{
\dontrun{
inat <- clean_inat_data(".\\\\data\\\\observations-17205.csv")
}
}
\seealso{
Other data preparation: \code{\link{NPS_data}},
  \code{\link{clean_NPS_syn}}, \code{\link{get_NPS_list}},
  \code{\link{inat_summary}}, \code{\link{taxataxi}}
}
\concept{data preparation}
