% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NPS_syn.R
\name{clean_NPS_syn}
\alias{clean_NPS_syn}
\title{Clean NPSpecies synonyms list}
\usage{
clean_NPS_syn(npssyn = NULL)
}
\arguments{
\item{npssyn}{Raw list of NPSpecies synonyms returned by \code{get_NPS_list}}
}
\value{
list of clean synonyms in a dat frame
}
\description{
Clean NPSpecies synonyms list
}
\examples{
\dontrun{
clean_NPS_syn(npssyn)
}
}
\seealso{
Other data preparation: \code{\link{NPS_data}},
  \code{\link{clean_inat_data}},
  \code{\link{get_NPS_list}}, \code{\link{inat_summary}},
  \code{\link{taxataxi}}
}
\concept{data preparation}
