% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NPS_list.R
\name{get_NPS_list}
\alias{get_NPS_list}
\title{Get NPSpecies data for single park}
\usage{
get_NPS_list(unitcode = "ABLI", listtype = "fulllist")
}
\arguments{
\item{unitcode}{Four letter Unit code for the park}

\item{listtype}{Type of list to be downloaded. Accepted values are "fulllist" for
for complete list, "checklist" for checklist of species, "detaillist" for
full list of species with details}
}
\value{
list of two data frames first \code{data} containing list of species and
 second \code{Syn} containing list of synonyms with respective accepted names
}
\description{
Get NPSpecies data for single park
}
\examples{
\dontrun{
getNPSlist(unitcode="ABLI", listtype="fulllist")
}
}
\seealso{
Other data preparation: \code{\link{NPS_data}},
  \code{\link{clean_NPS_syn}},
  \code{\link{clean_inat_data}},
  \code{\link{inat_summary}}, \code{\link{taxataxi}}
}
\concept{data preparation}
