% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_summary.R
\name{inat_summary}
\alias{inat_summary}
\title{Build a iNaturalist data summary table with Park codes and counts of records}
\usage{
inat_summary(inat = NULL)
}
\arguments{
\item{inat}{Clean iNaturalist data typically returned from \code{clean_inat_data}}
}
\value{
data frame containing iNaturalist data summary with \code{unit_code}
and \code{count}
}
\description{
Build a iNaturalist data summary table with Park codes and counts of records
}
\examples{
\dontrun{
inatsum <- inat_summary(inat)
}
}
\seealso{
Other data preparation: \code{\link{NPS_data}},
  \code{\link{clean_NPS_syn}},
  \code{\link{clean_inat_data}}, \code{\link{get_NPS_list}}
}
\concept{data preparation}
