% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_data.R
\name{NPS_data}
\alias{NPS_data}
\title{Download all individual Park checklist files}
\usage{
NPS_data(inatsum = NULL, outpath = NA)
}
\arguments{
\item{inatsum}{iNaturalist data summary typically returned from \code{inat_summary}}

\item{outpath}{Output path to store the NPSpecies files and results}
}
\value{
data frame containing iNaturalist data summary with \code{unit_code}
and \code{count}
}
\description{
Script to download individual Park checklist files
}
\examples{
\dontrun{
NPS_data(inatsum)
}
}
\seealso{
Other data preparation: \code{\link{clean_NPS_syn}},
  \code{\link{clean_inat_data}},
  \code{\link{get_NPS_list}}, \code{\link{inat_summary}}
}
\concept{data preparation}
