% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxataxi.R
\name{taxataxi}
\alias{taxataxi}
\title{main function}
\usage{
taxataxi(inat, NPSall, unitcode, rg = FALSE)
}
\arguments{
\item{inat}{iNaturalist data}

\item{NPSall}{NPS data for all Parks}

\item{unitcode}{unitcode of the park}

\item{rg}{If the function should consider only research grade
records (default FALSE)}
}
\value{
list of data frames consisting of summery tables, matching and
non matching records with NPS and ITIS names etc.
}
\description{
Script to
}
\examples{
\dontrun{
taxataxi("ABLI")
}
}
\seealso{
Other data preparation: \code{\link{NPS_data}},
  \code{\link{clean_NPS_syn}},
  \code{\link{clean_inat_data}},
  \code{\link{get_NPS_list}}, \code{\link{inat_summary}}
}
\concept{data preparation}
